requirements.txt
django
git+git://github.com/chase-seibert/django-social-auth.git@master#egg=social_auth==0.7.18
django-celery
feedparser
python-dateutil
BeautifulSoup
Pygments
bleach
dj-database-url
psycopg2
gunicorn
gevent
south
django-crispy-forms
django-colorful
